Report added

+++ src/Main.java
@@ -2,9 +2,9 @@ public class Main {
 
     public static void main(String[] args) {
 
-        long balans = 2_000_000_000;
+        int balans = 2_000_000_000;
         int amount = 500_000_000;
-        long total = balans + amount;
+        int total = balans + amount;
 
         System.out.println(total);
